Added Credit Card, Required Table in migration, Added Swagger, Deployed to Azure Portal

+++ Orien.PYS.Web/Orien.PYS.Data/Entity/SplitRelation.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         
-        public int Slip_Id { get; set; }
+        public long Slip_Id { get; set; }
         
         public int UserId { get; set; }
         public User User { get; set; }
